+++ RELEASE-NOTES.md
@@ -37,16 +37,18 @@ END TEMPLATE-->
 
 ### Breaking changes
 
-*None yet*
+* SpawnNextToOrDrop methods have been changed to take `EntProtoId?, Entity<TransformComponent?>, Vector2 = default` instead of `string?, EntityUid, TransformComponent? null`. Because of this, some `Entity<T>`s will need to first be explicity converted into `EntityUid`s.
+* DropNextTo and anything using it will now inherit the rotation of the target instead of the target's parent.
 
 ### New features
 
 * `Control.Restyle` has been made public for manually restyling controls on edge cases.
 * `EntityManager.PredictedSpawn(string?,ComponentRegistry?,bool)`, `EntityManager.PredictedSpawn(string?,MapCoordinates,ComponentRegistry?,Angle)`, `EntityManager.PredictedSpawnAtPosition(string?,EntityCoordinates,Angle,ComponentRegistry?)`, and `EntityManager.PredictedSpawnInContainerOrDrop(string?,EntityUid,string,out bool,TransformComponent?,ContainerManagerComponent?,ComponentRegistry?)` now all have proxies for EntitySystems.
+* SpawnNextToOrDrop methods can now offset their spawns by passing in a Vector2. This offset is always relative to the target.
 
 ### Bugfixes
 
-*None yet*
+* SpawnNextToOrDrop no longer runs mapinit on entities in nullspace.
 
 ### Other